backend/requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
pandas==1.3.3
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
flask-cors==5.0.1
XlsxWriter==3.2.2
gunicorn==20.1.0
requests==2.32.3
extract-msg==0.45.0
Werkzeug==3.1.3
itsdangerous==2.2.0